+++ CHANGELOG.md
@@ -5,9 +5,18 @@ This change log follows the conventions of [keepachangelog.com](http://keepachan
 
 ## [Unreleased]
 
+## [0.34.0] - 2022-07-17
+### Added
+-  Rate limiting for PyPI's public API requests with `--rate-limits
+   requests/milliseconds` CLI option. **NB!** Default rate limits are
+   set to 120 requests per minute. This release **may slow down** your
+   Python dependencies check significantly in case large number of
+   packages
+   ([#101](https://github.com/pilosus/pip-license-checker/issues/101))
+
 ## [0.33.0] - 2022-06-25
 ### Fixed
-- Parsing long numbers in Python versions ((#99)[https://github.com/pilosus/pip-license-checker/issues/99])
+- Parsing long numbers in Python versions ([#99](https://github.com/pilosus/pip-license-checker/issues/99))
 
 ## [0.32.0] - 2022-06-23
 ### Added
@@ -271,7 +280,8 @@ weak copyleft types.
 ### Added
 - Structure for Leiningen app project
 
-[Unreleased]: https://github.com/pilosus/pip-license-checker/compare/0.33.0...HEAD
+[Unreleased]: https://github.com/pilosus/pip-license-checker/compare/0.34.0...HEAD
+[0.34.0]: https://github.com/pilosus/pip-license-checker/compare/0.33.0...0.34.0
 [0.33.0]: https://github.com/pilosus/pip-license-checker/compare/0.32.0...0.33.0
 [0.32.0]: https://github.com/pilosus/pip-license-checker/compare/0.31.0...0.32.0
 [0.31.0]: https://github.com/pilosus/pip-license-checker/compare/0.30.0...0.31.0